♻️ refactor: extract wrapWithRTK into plugin hook rtk
Move RTK command rewriting from hardcoded BashTool.Execute() into
plugins/hooks/rtk/ as a PreToolCallHook plugin. The hook only fires
for bash tool calls and is enabled by default for backward compat.
